util: move UUID to an separate file
